Tasks-Android: Refactores tasks module & adds unit tests

+++ tasks-android/app/src/main/java/com/lotic/tasks/domain/modules/tasks/operations/tasks/ClearTasksForAccount.kt
@@ -1,7 +1,7 @@
 package com.lotic.tasks.domain.modules.tasks.operations.tasks
 
-import com.lotic.tasks.domain.modules.tasks.dtos.Task
-import com.lotic.tasks.domain.modules.tasks.repositories.TasksRepository
+import com.lotic.tasks.domain.modules.tasks.Task
+import com.lotic.tasks.domain.modules.tasks.TasksRepository
 import com.lotic.tasks.domain.shared.Command
 import java.util.*
 
@@ -9,7 +9,9 @@ class ClearTasksForAccount(private val tasksRepository: TasksRepository) : Comma
 
     override suspend fun execute(input: UUID) {
         val taskList: List<Task> = this.tasksRepository.listTasksForAccount(input)
-        this.tasksRepository.deleteMultiple(taskList.map { it.id })
+        if (taskList.isNotEmpty()) {
+            this.tasksRepository.deleteMultiple(taskList.map { it.id })
+        }
     }
 
 }
